Moved preprocessing stage of mfiles importer to class.
